requirements.txt
customtkinter
Pillow
pywin32
numpy
pyinstaller
pyttsx3
opencv-python-headless
openai
